+++ src/pages/lancer-un-defi/linguistique/3.tsx
@@ -1,9 +1,5 @@
 import { useRouter } from 'next/router';
-import { useSnackbar } from 'notistack';
 import React from 'react';
-
-import type { SelectChangeEvent } from '@mui/material';
-import { FormControl, Select, MenuItem } from '@mui/material';
 
 import { isDefi } from 'src/activity-types/anyActivity';
 import { isLanguage, LANGUAGE_OBJECTS } from 'src/activity-types/defi.constants';
@@ -12,58 +8,66 @@ import { Base } from 'src/components/Base';
 import { StepsButton } from 'src/components/StepsButtons';
 import { Steps } from 'src/components/Steps';
 import { ContentEditor } from 'src/components/activities/content';
+import { getErrorSteps } from 'src/components/activities/defiLanguageChecks';
 import { ActivityContext } from 'src/contexts/activityContext';
-import { replaceTokens } from 'src/utils';
 import type { ActivityContent, ActivityContentType } from 'types/activity.type';
-import { ActivityStatus } from 'types/activity.type';
 
 const DefiStep3 = () => {
   const router = useRouter();
-  const { enqueueSnackbar } = useSnackbar();
   const { activity, updateActivity, addContent, deleteContent, save } = React.useContext(ActivityContext);
 
   const data = (activity?.data as LanguageDefiData) || null;
   const explanationContentIndex = Math.max(data?.explanationContentIndex ?? 0, 0);
-  const isEdit = activity !== null && activity.id !== 0 && activity.status !== ActivityStatus.DRAFT;
 
+  const errorSteps = React.useMemo(() => {
+    if (data !== null) {
+      return getErrorSteps(data, 2);
+    }
+    return [];
+  }, [data]);
+
+  const contentAdded = React.useRef(false);
   React.useEffect(() => {
     if (activity === null && !('activity-id' in router.query) && !sessionStorage.getItem('activity')) {
       router.push('/lancer-un-defi');
     } else if (activity && (!isDefi(activity) || (isDefi(activity) && !isLanguage(activity)))) {
       router.push('/lancer-un-defi');
     }
-  }, [activity, router]);
+
+    if (activity && isDefi(activity) && isLanguage(activity)) {
+      if ((activity.data.explanationContentIndex ?? 0) > activity.content.length) {
+        updateActivity({
+          data: {
+            ...activity.data,
+            explanationContentIndex: activity.content.length,
+          },
+        });
+      }
+      if ((activity.data.explanationContentIndex ?? 0) === activity.content.length && !contentAdded.current) {
+        contentAdded.current = true;
+        addContent('text');
+      }
+    }
+  }, [activity, router, updateActivity, addContent]);
 
   if (data === null || activity === null || !isDefi(activity) || (isDefi(activity) && !isLanguage(activity))) {
     return <div></div>;
   }
 
   const updateContent = (content: ActivityContent[]): void => {
-    updateActivity({ content: [...content, ...activity.content.slice(explanationContentIndex, activity.content.length)] });
+    updateActivity({ content: [...activity.content.slice(0, explanationContentIndex), ...content] });
   };
-  const addDescriptionContent = (type: ActivityContentType, value?: string) => {
-    addContent(type, value, explanationContentIndex);
-    updateActivity({ data: { ...data, explanationContentIndex: explanationContentIndex + 1 } });
+  const addIndiceContent = (type: ActivityContentType, value?: string) => {
+    contentAdded.current = true;
+    addContent(type, value);
   };
-  const deleteDescriptionContent = (index: number) => {
-    deleteContent(index);
-    updateActivity({ data: { ...data, explanationContentIndex: explanationContentIndex - 1 } });
-  };
-
-  const onObjectChange = (event: SelectChangeEvent<string | number>) => {
-    updateActivity({ data: { ...data, objectIndex: `${event.target.value}` } });
+  const deleteIndiceContent = (index: number) => {
+    contentAdded.current = true; // delete means there were content already
+    deleteContent(explanationContentIndex + index);
   };
 
   const onNext = () => {
-    if (data.objectIndex === -1) {
-      return;
-    }
-    if (explanationContentIndex === 0) {
-      enqueueSnackbar('Il faut au moins un bloc de texte, image, son ou vidéo avant de continuer.', {
-        variant: 'error',
-      });
-      return;
-    }
+    save().catch(console.error);
     router.push('/lancer-un-defi/linguistique/4');
   };
 
@@ -71,45 +75,29 @@ const DefiStep3 = () => {
     <Base>
       <div style={{ width: '100%', padding: '0.5rem 1rem 1rem 1rem' }}>
         <Steps
-          steps={(isEdit ? [] : ['Démarrer']).concat(['Choix de la langue', "Choix de l'objet", 'Explication', 'Le défi', 'Prévisualisation'])}
-          activeStep={isEdit ? 1 : 2}
+          steps={['Choix de la langue', "Choix de l'objet", 'Explication', 'Le défi', 'Prévisualisation']}
+          urls={[
+            '/lancer-un-defi/linguistique/1?edit',
+            '/lancer-un-defi/linguistique/2',
+            '/lancer-un-defi/linguistique/3',
+            '/lancer-un-defi/linguistique/4',
+            '/lancer-un-defi/linguistique/5',
+          ]}
+          activeStep={2}
+          errorSteps={errorSteps}
         />
         <div className="width-900">
-          <h1>{"Choix de l'objet"}</h1>
+          <h1>{'Explication'}</h1>
           <p className="text" style={{ fontSize: '1.1rem' }}>
-            Choisissez ce que vous souhaitez présenter :
+            {LANGUAGE_OBJECTS[data?.objectIndex % LANGUAGE_OBJECTS.length]?.desc2}
           </p>
-          <FormControl variant="outlined" style={{ width: '100%' }}>
-            <Select
-              labelId="demo-simple-select-outlined-label"
-              id="demo-simple-select-outlined"
-              value={data.objectIndex === -1 ? '' : data.objectIndex}
-              onChange={onObjectChange}
-            >
-              {LANGUAGE_OBJECTS.map((l, index) => (
-                <MenuItem key={index} value={index}>
-                  {l.title}
-                </MenuItem>
-              ))}
-            </Select>
-          </FormControl>
-          {data.objectIndex !== -1 && (
-            <>
-              <p className="text" style={{ fontSize: '1.1rem' }}>
-                {replaceTokens(LANGUAGE_OBJECTS[data.objectIndex % LANGUAGE_OBJECTS.length].desc1, {
-                  language: data.language,
-                })}{' '}
-                Vous pouvez rajouter une vidéo ou un son pour qu’on entende la prononciation.
-              </p>
-              <ContentEditor
-                content={activity.content.slice(0, explanationContentIndex)}
-                updateContent={updateContent}
-                addContent={addDescriptionContent}
-                deleteContent={deleteDescriptionContent}
-                save={save}
-              />
-            </>
-          )}
+          <ContentEditor
+            content={activity.content.slice(explanationContentIndex, activity.content.length)}
+            updateContent={updateContent}
+            addContent={addIndiceContent}
+            deleteContent={deleteIndiceContent}
+            save={save}
+          />
           <StepsButton prev="/lancer-un-defi/linguistique/2" next={onNext} />
         </div>
       </div>
