+++ src/pages/lancer-un-defi/linguistique/5.tsx
@@ -1,35 +1,48 @@
+import classNames from 'classnames';
+import Link from 'next/link';
 import { useRouter } from 'next/router';
 import React from 'react';
 
-import { TextField, Button } from '@mui/material';
+import Backdrop from '@mui/material/Backdrop';
+import Button from '@mui/material/Button';
+import CircularProgress from '@mui/material/CircularProgress';
 
 import { isDefi } from 'src/activity-types/anyActivity';
-import { isLanguage, LANGUAGE_DEFIS, LANGUAGE_OBJECTS } from 'src/activity-types/defi.constants';
+import { isLanguage, getDefi, getLanguageObject, DEFI } from 'src/activity-types/defi.constants';
 import type { LanguageDefiData } from 'src/activity-types/defi.types';
 import { Base } from 'src/components/Base';
 import { StepsButton } from 'src/components/StepsButtons';
 import { Steps } from 'src/components/Steps';
-import { ThemeChoiceButton } from 'src/components/buttons/ThemeChoiceButton';
+import ActivityLink from 'src/components/activities/Link';
+import { Activities } from 'src/components/activities/List';
+import { ContentView } from 'src/components/activities/content/ContentView';
+import { getErrorSteps } from 'src/components/activities/defiLanguageChecks';
+import { REACTIONS } from 'src/components/activities/utils';
+import { EditButton } from 'src/components/buttons/EditButton';
 import { ActivityContext } from 'src/contexts/activityContext';
-import { replaceTokens } from 'src/utils';
+import { useActivity } from 'src/services/useActivity';
 import { ActivityStatus } from 'types/activity.type';
 
-const DefiStep6 = () => {
+const DefiStep5 = () => {
   const router = useRouter();
-  const { activity, updateActivity } = React.useContext(ActivityContext);
-  const [otherOpen, setIsOtherOpen] = React.useState(false);
+  const { activity, save } = React.useContext(ActivityContext);
+  const { activity: responseActivity } = useActivity(activity?.responseActivityId ?? -1);
+  const [isLoading, setIsLoading] = React.useState(false);
 
   const data = (activity?.data as LanguageDefiData) || null;
+  const explanationContentIndex = Math.max(data?.explanationContentIndex ?? 0, 0);
   const isEdit = activity !== null && activity.id !== 0 && activity.status !== ActivityStatus.DRAFT;
 
-  const c = data?.defi || '';
-  const opened = React.useRef(false);
-  React.useEffect(() => {
-    if (c && !opened.current) {
-      setIsOtherOpen(true);
-      opened.current = true;
+  const errorSteps = React.useMemo(() => {
+    if (activity?.data.explanationContentIndex === activity?.content.length && data !== null) {
+      const errors = getErrorSteps(data, 3);
+      errors.push(2); //corresponding to step 3
+      return errors;
     }
-  }, [c]);
+    if (data !== null) return getErrorSteps(data, 3);
+    return [];
+  }, [activity?.content.length, activity?.data.explanationContentIndex, data]);
+  const isValid = errorSteps.length === 0;
 
   React.useEffect(() => {
     if (activity === null && !('activity-id' in router.query) && !sessionStorage.getItem('activity')) {
@@ -39,88 +52,146 @@ const DefiStep6 = () => {
     }
   }, [activity, router]);
 
+  const onPublish = async () => {
+    setIsLoading(true);
+    const success = await save(true);
+    if (success) {
+      router.push('/lancer-un-defi/success');
+    }
+    setIsLoading(false);
+  };
+
   if (data === null || activity === null || !isDefi(activity) || (isDefi(activity) && !isLanguage(activity))) {
     return <div></div>;
   }
 
-  const onClick = (index: number) => () => {
-    if (index === -1) {
-      if (!data.defi) {
-        return;
-      }
-      updateActivity({ data: { ...data, defiIndex: index, defi: data.defi.toLowerCase() } });
-    } else {
-      const newData = data;
-      delete newData.defi;
-      updateActivity({ data: { ...newData, defiIndex: index } });
-    }
-    router.push('/lancer-un-defi/linguistique/6');
-  };
-
   return (
     <Base>
       <div style={{ width: '100%', padding: '0.5rem 1rem 1rem 1rem' }}>
-        <Steps
+        {/* <Steps
           steps={(isEdit ? [] : ['Démarrer']).concat(['Choix de la langue', "Choix de l'objet", 'Explication', 'Le défi', 'Prévisualisation'])}
-          activeStep={isEdit ? 3 : 4}
+          activeStep={isEdit ? 4 : 5}
+        /> */}
+        <Steps
+          steps={['Choix de la langue', "Choix de l'objet", 'Explication', 'Le défi', 'Prévisualisation']}
+          urls={[
+            '/lancer-un-defi/linguistique/1?edit',
+            '/lancer-un-defi/linguistique/2',
+            '/lancer-un-defi/linguistique/3',
+            '/lancer-un-defi/linguistique/4',
+            '/lancer-un-defi/linguistique/5',
+          ]}
+          activeStep={4}
+          errorSteps={errorSteps}
         />
         <div className="width-900">
-          <h1>Quel défi voulez-vous lancer aux Pelicopains ?</h1>
-          <div style={{ marginTop: '1rem' }}>
-            {LANGUAGE_DEFIS.map((t, index) => (
-              <ThemeChoiceButton
-                key={index}
-                label={
-                  data.objectIndex === 4 && index === 0
-                    ? 'Trouvez la même chose dans une autre langue'
-                    : replaceTokens(t.title, {
-                        object:
-                          index === 0
-                            ? LANGUAGE_OBJECTS[data.objectIndex % LANGUAGE_OBJECTS.length].title.toLowerCase()
-                            : LANGUAGE_OBJECTS[data.objectIndex % LANGUAGE_OBJECTS.length].title2,
-                        language: data.language,
-                      })
-                }
-                description={t.description}
-                onClick={onClick(index)}
-              />
-            ))}
-            <ThemeChoiceButton
-              isOpen={otherOpen}
+          <h1>Pré-visualisez votre défi{!isEdit && ', et publiez-la'}</h1>
+          <p className="text" style={{ fontSize: '1.1rem' }}>
+            Voici la pré-visualisation de votre défi.
+            {isEdit
+              ? " Vous pouvez le modifier à l'étape précédente, et enregistrer vos changements ici."
+              : ' Vous pouvez le modifier, et quand vous êtes prêts : publiez-le dans votre village-monde !'}
+          </p>
+          {isEdit ? (
+            <div style={{ width: '100%', display: 'flex', justifyContent: 'space-between', margin: '1rem 0' }}>
+              <Link href="/lancer-un-defi/linguistique/5" passHref>
+                <Button component="a" color="secondary" variant="contained" href="/lancer-un-defi/linguistique/5">
+                  {"Modifier à l'étape précédente"}
+                </Button>
+              </Link>
+              <Button variant="outlined" color="primary" onClick={onPublish}>
+                Enregistrer les changements
+              </Button>
+            </div>
+          ) : (
+            <div style={{ width: '100%', textAlign: 'right', margin: '1rem 0' }}>
+              <Button variant="outlined" color="primary" onClick={onPublish}>
+                Publier
+              </Button>
+            </div>
+          )}
+
+          {!isEdit && activity.responseActivityId === null && <ActivityLink url={`/lancer-un-defi/linguistique/1?edit=${activity.id}`} />}
+
+          {responseActivity !== null && (
+            <>
+              <span className="text text--small text--success">Défi en réaction à {REACTIONS[responseActivity.type]}</span>
+              <div className="preview-block">
+                {!isEdit && (
+                  <EditButton
+                    onClick={() => {
+                      router.push(`/lancer-un-defi/linguistique/1?edit=${activity.id}`);
+                    }}
+                    status={'success'}
+                    style={{ position: 'absolute', top: '0.5rem', right: '0.5rem' }}
+                  />
+                )}
+                <Activities activities={[responseActivity]} noButtons />
+              </div>
+            </>
+          )}
+
+          <div className={classNames('preview-block', { 'preview-block--warning': !isValid && errorSteps.includes(0) })}>
+            <EditButton
               onClick={() => {
-                setIsOtherOpen(!otherOpen);
+                router.push('/lancer-un-defi/linguistique/1');
               }}
-              additionalContent={
-                <div>
-                  <div style={{ display: 'flex', alignItems: 'center', margin: '0.5rem 0' }}>
-                    <span style={{ marginRight: '0.3rem' }}>Défi : </span>
-                    {data !== null && (
-                      <TextField
-                        variant="standard"
-                        value={data.defi || ''}
-                        onChange={(event) => {
-                          updateActivity({ data: { ...data, defi: event.target.value } });
-                        }}
-                        style={{ minWidth: '0', flex: 1 }}
-                      />
-                    )}
-                  </div>
-                  <div className="text-center" style={{ marginTop: '0.8rem' }}>
-                    <Button color="primary" size="small" variant="outlined" onClick={onClick(-1)}>
-                      Continuer
-                    </Button>
-                  </div>
-                </div>
-              }
-              label="Un autre défi"
-              description={`Rédigez vous même le défi pour vos Pelicopains !`}
+              status={errorSteps.includes(0) ? 'warning' : 'success'}
+              style={{ position: 'absolute', top: '0.5rem', right: '0.5rem' }}
             />
+            {getLanguageObject(data)}
           </div>
+
+          <span className={classNames('text text--small text--success', { 'text text--small text--warning': !isValid && errorSteps.includes(1) })}>
+            {"L'expression"}
+          </span>
+          <div className={classNames('preview-block', { 'preview-block--warning': !isValid && errorSteps.includes(1) })}>
+            <EditButton
+              onClick={() => {
+                router.push('/lancer-un-defi/linguistique/2');
+              }}
+              status={errorSteps.includes(1) ? 'warning' : 'success'}
+              style={{ position: 'absolute', top: '0.5rem', right: '0.5rem' }}
+            />
+            <ContentView content={activity.content.slice(0, explanationContentIndex)} />
+          </div>
+
+          <span className={classNames('text text--small text--success', { 'text text--small text--warning': !isValid && errorSteps.includes(2) })}>
+            Explication
+          </span>
+          <div className={classNames('preview-block', { 'preview-block--warning': !isValid && errorSteps.includes(2) })}>
+            <EditButton
+              onClick={() => {
+                router.push('/lancer-un-defi/linguistique/3');
+              }}
+              status={errorSteps.includes(2) ? 'warning' : 'success'}
+              style={{ position: 'absolute', top: '0.5rem', right: '0.5rem' }}
+            />
+            <ContentView content={activity.content.slice(explanationContentIndex, activity.content.length)} />
+          </div>
+
+          <span className={classNames('text text--small text--success', { 'text text--small text--warning': !isValid && errorSteps.includes(3) })}>
+            Le défi lancé aux Pélicopains
+          </span>
+          <div className={classNames('preview-block', { 'preview-block--warning': !isValid && errorSteps.includes(3) })}>
+            <EditButton
+              onClick={() => {
+                router.push('/lancer-un-defi/linguistique/4');
+              }}
+              status={errorSteps.includes(3) ? 'warning' : 'success'}
+              style={{ position: 'absolute', top: '0.5rem', right: '0.5rem' }}
+            />
+            Votre défi : {getDefi(DEFI.LANGUAGE, data)}
+          </div>
+
           <StepsButton prev="/lancer-un-defi/linguistique/4" />
         </div>
       </div>
+      <Backdrop style={{ zIndex: 2000, color: 'white' }} open={isLoading}>
+        <CircularProgress color="inherit" />
+      </Backdrop>
     </Base>
   );
 };
 
-export default DefiStep6;
+export default DefiStep5;
